add: connection to api
