+++ frontend/src/components/App/App.js
@@ -5,7 +5,7 @@ import Home from "components/pages/Home";
 import Videos from "components/pages/Videos";
 import Profile from "components/pages/Profile";
 import Navigation from "components/Navigation";
-import MentorGallery from "components/pages/MentorGallery";
+import Gallery from "components/pages/Gallery";
 
 function App() {
   return (
@@ -43,7 +43,7 @@ function App() {
         path="/gallery"
         component={() => (
           <Navigation
-            content={<MentorGallery />}
+            content={<Gallery />}
             page="gallery"
             needsAuth={false}
           />
